initialize game logic
